+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/Settings.kt
@@ -340,6 +340,11 @@ abstract class Settings {
      * in TLS and HTTP/3.
      */
     open var postQuantumKeyExchangeEnabled: Boolean by UnsupportedSetting()
+
+    /**
+     * Comma-separated list of destination ports that the application should block connections to.
+     */
+    open var bannedPorts: String by UnsupportedSetting()
 }
 
 /**
@@ -408,6 +413,7 @@ data class DefaultSettings(
     override var cookieBehaviorOptInPartitioningPBM: Boolean = false,
     override var certificateTransparencyMode: Int = 0,
     override var postQuantumKeyExchangeEnabled: Boolean = false,
+    override var bannedPorts: String = "",
 ) : Settings() {
     override val desktopModeEnabled: Boolean
         get() = getDesktopMode()