+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/components/Core.kt
@@ -190,6 +190,7 @@ class Core(
             parallelMarkingEnabled = FxNimbus.features.javascript.value().parallelMarkingEnabled,
             certificateTransparencyMode = FxNimbus.features.pki.value().certificateTransparencyMode,
             postQuantumKeyExchangeEnabled = FxNimbus.features.pqcrypto.value().postQuantumKeyExchangeEnabled,
+            bannedPorts = FxNimbus.features.networkingBannedPorts.value().bannedPortList,
         )
 
         // Apply fingerprinting protection overrides if the feature is enabled in Nimbus